Add hints for O(1) lookup of some read indexes
This is the beginning of code that will reduce the amount of searching
needed to find a given index with rafter_log:get_entry. It will keep a
dictionary of previously seen index locations in the logfile and use
that as a cache for rafter_log:find_entry calls that search the log. As
it is very expensive to store pointers to all indexes, this hint
strategy will attempt to be smart and only store a subset of frequently
needed pointers.

Basic stastics are kept to keep track of how efficient the
implementation is.

It is anticipated in the future that search will find the "closest" hint
and not just need an exact cache hit. This should greatly reduce seeks
as well.

It is also possible in the future that hints will be persisted to their
individual log section files.