Fix log corruption bug during leader overwrite
Fixes the issue discovered here:
https://github.com/andrewjstone/rafter/pull/32

When an existing, but uncommitted entry gets overwritten by a newly
elected leader, the index in the new entry is incorrectly set. In the
example given:

Follower has an entry with term=1, index=1
Leader sends new entry with term=2, index=1
This should lead to the entry in the follower's log containing term=1,
index1, however before the bug was fixed the new entry was term=1,
index=2.

This fixes this error with some refactoring and cleanup. Now all log
entries coming from a leader and being appended by the follower are
appended exactly as sent. Log indexes are only directly incremented when
a leader is appending a new operation to its own log and hence doesn't
have an index in the entry being appended. In this case the log state is
used. As a result, the log state is also correctly updated in the log
process.

While working on this issue, another bug was found. If the leader
overwrites a configuration entry, the `config_loc` and`config` state
members need to be updated with the last valid configuration (the one in
the trailer of the entry immediately before the truncation point). This
was fixed as well. It was also discovered during testing that the `term`
member of the log state is unused. In order to prevent confusion it was
removed from the state record.

A test covering these changes was added as well.